Replace ConditionalExpectationConfigurationBuilder with DefaultExpectationConfigurationBuilder
