+++ great_expectations/rule_based_profiler/expectation_configuration_builder/__init__.py
@@ -5,7 +5,3 @@ from great_expectations.rule_based_profiler.expectation_configuration_builder.ex
 from great_expectations.rule_based_profiler.expectation_configuration_builder.default_expectation_configuration_builder import (  # isort:skip
     DefaultExpectationConfigurationBuilder,
 )
-
-from great_expectations.rule_based_profiler.expectation_configuration_builder.conditional_expectation_configuration_builder import (
-    ConditionalExpectationConfigurationBuilder,
-)
