+++ tests/rule_based_profiler/expectation_configuration_builder/test_default_expectation_configuration_builder.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 import pytest
 
 import great_expectations.exceptions as ge_exceptions
@@ -13,6 +15,7 @@ from great_expectations.rule_based_profiler.parameter_builder import (
 from great_expectations.rule_based_profiler.types import (
     Domain,
     ParameterContainer,
+    build_parameter_container_for_variables,
     get_parameter_value_by_fully_qualified_parameter_name,
 )
 
@@ -56,16 +59,16 @@ def test_meta_not_dict_exception(
         parameters={domain.id: parameter_container},
     )
 
-    max_user_id = 999999999999
+    condition = None
+    max_user_id: int = 999999999999
 
     with pytest.raises(ge_exceptions.ProfilerExecutionError) as e:
-        default_expectation_configuration_builder = (
-            DefaultExpectationConfigurationBuilder(
-                expectation_type="expect_column_values_to_be_between",
-                min_value=value.value[0],
-                max_value=max_user_id,
-                meta="Strings are not acceptable",
-            )
+        DefaultExpectationConfigurationBuilder(
+            expectation_type="expect_column_values_to_be_between",
+            condition=condition,
+            min_value=value.value[0],
+            max_value=max_user_id,
+            meta="Strings are not acceptable",
         )
 
     assert (
@@ -74,7 +77,7 @@ def test_meta_not_dict_exception(
     )
 
 
-def test_default_expectation_configuration_builder_alice(
+def test_condition_not_string_exception(
     alice_columnar_table_single_batch_context,
 ):
     data_context: DataContext = alice_columnar_table_single_batch_context
@@ -113,18 +116,811 @@ def test_default_expectation_configuration_builder_alice(
         parameters={domain.id: parameter_container},
     )
 
-    max_user_id = 999999999999
+    condition: dict = {"condition": "$variables.tolerance<0.8"}
+    max_user_id: int = 999999999999
 
-    default_expectation_configuration_builder = DefaultExpectationConfigurationBuilder(
+    with pytest.raises(ge_exceptions.ProfilerExecutionError) as e:
+        DefaultExpectationConfigurationBuilder(
+            expectation_type="expect_column_values_to_be_between",
+            condition=condition,
+            min_value=value.value[0],
+            max_value=max_user_id,
+        )
+
+    assert (
+        str(e.value)
+        == 'Argument "{\'condition\': \'$variables.tolerance<0.8\'}" in "DefaultExpectationConfigurationBuilder" must be of type "string" (value of type "<class \'dict\'>" was encountered).\n'
+    )
+
+
+def test_default_expectation_configuration_builder_alice_null_condition(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id"
+    value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition = None
+    max_user_id: int = 999999999999
+
+    with pytest.raises(ge_exceptions.ProfilerExecutionError) as e:
+        DefaultExpectationConfigurationBuilder(
+            expectation_type="expect_column_values_to_be_between",
+            condition=condition,
+            min_value=value.value[0],
+            max_value=max_user_id,
+        )
+
+    assert (
+        str(e.value)
+        == 'Argument "None" in "DefaultExpectationConfigurationBuilder" must be of type "string" (value of type "<class \'NoneType\'>" was encountered).\n'
+    )
+
+
+def test_default_expectation_configuration_builder_alice_single_term_parameter_condition_true(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id"
+    value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$parameter.my_min_user_id.value[0]>0"
+    max_user_id: int = 999999999999
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
         expectation_type="expect_column_values_to_be_between",
+        condition=condition,
         min_value=value.value[0],
         max_value=max_user_id,
     )
 
-    expectation_configuration: ExpectationConfiguration = (
-        default_expectation_configuration_builder._build_expectation_configuration(
-            domain=domain, parameters={domain.id: parameter_container}
-        )
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}
     )
 
     assert expectation_configuration.kwargs["min_value"] == 397433
+
+
+def test_default_expectation_configuration_builder_alice_single_term_parameter_condition_false(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id"
+    value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$parameter.my_min_user_id.value[0]<0"
+    max_user_id: int = 999999999999
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=value.value[0],
+        max_value=max_user_id,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}
+    )
+
+    assert expectation_configuration is None
+
+
+def test_default_expectation_configuration_builder_alice_single_term_variable_condition_true(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id>0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value.value[0],
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration.kwargs["min_value"] == 397433
+
+
+def test_default_expectation_configuration_builder_alice_single_term_variable_condition_false(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value.value[0],
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration is None
+
+
+def test_default_expectation_configuration_builder_alice_two_term_and_parameter_variable_condition_true(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id>0 & $parameter.my_min_user_id.value[0]>0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration.kwargs["min_value"] == 397433
+
+
+def test_default_expectation_configuration_builder_alice_two_term_and_parameter_variable_condition_false(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id<0 & $parameter.my_min_user_id.value[0]<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration is None
+
+
+def test_default_expectation_configuration_builder_alice_two_term_or_parameter_variable_condition_true(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id>0 | $parameter.my_min_user_id.value[0]<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration.kwargs["min_value"] == 397433
+
+
+def test_default_expectation_configuration_builder_alice_two_term_or_parameter_variable_condition_false(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id<0 | $parameter.my_min_user_id.value[0]<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration is None
+
+
+def test_default_expectation_configuration_builder_alice_more_than_two_term_parameter_variable_condition_true(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999, "answer": 42}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id>0 & $variables.answer==42 | $parameter.my_min_user_id.value[0]<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration.kwargs["min_value"] == 397433
+
+
+def test_default_expectation_configuration_builder_alice_more_than_two_term_parameter_variable_condition_false(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999, "answer": 42}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "$variables.max_user_id<0 | $variables.answer!=42 | $parameter.my_min_user_id.value[0]<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration is None
+
+
+def test_default_expectation_configuration_builder_alice_parentheses_parameter_variable_condition_true(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999, "answer": 42}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "($variables.max_user_id>0 & $variables.answer==42) | $parameter.my_min_user_id.value[0]<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration.kwargs["min_value"] == 397433
+
+
+def test_default_expectation_configuration_builder_alice_parentheses_parameter_variable_condition_false(
+    alice_columnar_table_single_batch_context,
+):
+    data_context: DataContext = alice_columnar_table_single_batch_context
+
+    batch_request: dict = {
+        "datasource_name": "alice_columnar_table_single_batch_datasource",
+        "data_connector_name": "alice_columnar_table_single_batch_data_connector",
+        "data_asset_name": "alice_columnar_table_single_batch_data_asset",
+    }
+
+    metric_domain_kwargs = {"column": "user_id"}
+
+    min_user_id_parameter: MetricMultiBatchParameterBuilder = (
+        MetricMultiBatchParameterBuilder(
+            name="my_min_user_id",
+            metric_name="column.min",
+            metric_domain_kwargs=metric_domain_kwargs,
+            data_context=data_context,
+            batch_request=batch_request,
+        )
+    )
+
+    parameter_container: ParameterContainer = ParameterContainer(parameter_nodes=None)
+    domain: Domain = Domain(
+        domain_type=MetricDomainTypes.COLUMN, domain_kwargs=metric_domain_kwargs
+    )
+
+    min_user_id_parameter._build_parameters(
+        parameter_container=parameter_container, domain=domain
+    )
+
+    variables: ParameterContainer = build_parameter_container_for_variables(
+        {"max_user_id": 999999999999, "answer": 42}
+    )
+
+    fully_qualified_parameter_name_for_value: str = "$parameter.my_min_user_id.value[0]"
+    parameter_value = get_parameter_value_by_fully_qualified_parameter_name(
+        fully_qualified_parameter_name=fully_qualified_parameter_name_for_value,
+        domain=domain,
+        parameters={domain.id: parameter_container},
+    )
+
+    condition: str = "($variables.max_user_id<0 | $variables.answer!=42) | $parameter.my_min_user_id.value[0]<0"
+    max_value: str = "$variables.max_user_id"
+
+    default_expectation_configuration_builder: DefaultExpectationConfigurationBuilder = DefaultExpectationConfigurationBuilder(
+        expectation_type="expect_column_values_to_be_between",
+        condition=condition,
+        min_value=parameter_value,
+        max_value=max_value,
+    )
+
+    expectation_configuration: Optional[
+        ExpectationConfiguration
+    ] = default_expectation_configuration_builder._build_expectation_configuration(
+        domain=domain, parameters={domain.id: parameter_container}, variables=variables
+    )
+
+    assert expectation_configuration is None
